requirements.txt
jupyter==1.0.0
setuptools==41.0.1
matplotlib==3.0.3
pandas==1.0.5
numpy==1.19.0
boto3==1.14.19
s3fs==0.4.2
PyMySQL==0.10.1
sklearn==0.0
Pillow==8.2.0
opyrator==0.0.8
tensorflow==2.2.0
opencv-python==4.5.1.48
